requirements.txt
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
idna==3.7
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.2
python-dotenv==1.0.1
requests==2.32.3
urllib3==2.2.1
